+++ tests/storeMiddlewares.test.ts
@@ -1,0 +1,140 @@
+import { createStore, createPiceOfStore } from '../src/core';
+import { IPushAction, IAction, IStatirMiddleware } from '../src/types';
+
+const piceOfState = createPiceOfStore(() => ({
+  state: {
+    count: 0
+  },
+  pipes: {
+    test: {
+      push(state: any, payload: number) {
+        return {
+          ...state,
+          count: payload
+        };
+      }
+    }
+  }
+}));
+
+describe('Testing store middlewares', () => {
+  test('invalid config', () => {
+    expect(() =>
+      createStore({
+        pices: { piceOfState },
+        middlewares: [(() => {}) as any]
+      })
+    ).not.toThrow();
+
+    expect(() =>
+      createStore({
+        pices: { piceOfState },
+        middlewares: [
+          (() => {
+            return () => {};
+          }) as IStatirMiddleware
+        ]
+      })
+    ).not.toThrow();
+
+    expect(() =>
+      createStore({
+        pices: { piceOfState },
+        middlewares: [
+          (next: IPushAction) => {
+            return (action: IAction) => {
+              next(action);
+            };
+          }
+        ]
+      })
+    ).not.toThrow();
+  });
+
+  test('state safety', () => {
+    const store = createStore({
+      pices: { piceOfState },
+      middlewares: [
+        (next: IPushAction) => {
+          return (action: IAction) => {
+            next(action);
+          };
+        }
+      ]
+    });
+
+    expect(store.state.piceOfState.count).toEqual(0);
+  });
+
+  test('dispatch safety', () => {
+    const store = createStore({
+      pices: { piceOfState },
+      middlewares: [
+        (next: IPushAction) => {
+          return (action: IAction) => {
+            next(action);
+          };
+        }
+      ]
+    });
+
+    expect(store.state.piceOfState.count).toEqual(0);
+
+    store.dispatch.piceOfState.test(123);
+
+    expect(store.state.piceOfState.count).toEqual(123);
+  });
+
+  test('listners safety', () => {
+    const listner = jest.fn(() => {});
+
+    const store = createStore({
+      pices: { piceOfState },
+      middlewares: [
+        (next: IPushAction) => {
+          return (action: IAction) => {
+            next(action);
+          };
+        }
+      ]
+    });
+
+    store.addListner(listner);
+
+    store.dispatch.piceOfState.test(123);
+
+    expect(listner.mock.calls.length).toEqual(2);
+  });
+
+  test('upgrade compatibility', () => {
+    const listner = jest.fn(() => {});
+
+    const store = createStore({
+      pices: { piceOfState },
+      middlewares: [
+        (next: IPushAction) => {
+          return (action: IAction) => {
+            next(action);
+          };
+        }
+      ],
+      upgrades: [
+        ((createStore: any) => {
+          return (config: any) => {
+            return createStore(config);
+          };
+        }) as any
+      ]
+    });
+
+    store.addListner(listner);
+
+    expect(store.state.piceOfState.count).toEqual(0);
+
+    store.dispatch.piceOfState.test(123);
+
+    expect(listner.mock.calls.length).toEqual(2);
+
+    expect(store.state.piceOfState.count).toEqual(123);
+  });
+});